pyproject.toml
[project]
name = "pimo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.3.0",
    "pytz>=2025.2",
    "requests>=2.32.4",
]
